+++ TIUIKitCore/Sources/Presenter/UIViewControllerPresenter.swift
@@ -20,24 +20,11 @@
 //  THE SOFTWARE.
 //
 
-import Moya
-import Foundation
+import UIKit.UIViewController
 
-public enum EndpointErrorResult<ApiError, NetworkError>: Error {
-    case apiError(ApiError)
-    case networkError(NetworkError)
-}
+@MainActor
+public protocol UIViewControllerPresenter {
+    associatedtype ViewController: UIViewController
 
-public extension EndpointErrorResult where NetworkError == MoyaError {
-    var isNetworkConnectionProblem: Bool {
-        guard case let .networkError(moyaError) = self,
-              case let .underlying(error, _) = moyaError,
-              case let .sessionTaskFailed(urlSessionTaskError) = error.asAFError,
-              let urlError = urlSessionTaskError as? URLError else {
-
-                  return false
-              }
-
-        return urlError.code == .notConnectedToInternet
-    }
+    func createViewController() -> ViewController
 }